så kan man modtage oprettede ressourser

+++ MJAPDelpin.Contract.Domain/DTO/DTORessource.cs
@@ -9,11 +9,11 @@ namespace MJAPDelpin.Contract.Domain.DTO
        public int RessourceId { get; set; }
         public string RessourceModelString { get; set; }
 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public bool RessourceState { get; set; }
 
 
-       public DTORessource(int ressourceid, string ressourceModelString, bool ressourceState, decimal price)
+       public DTORessource(int ressourceid, string ressourceModelString, bool ressourceState, int price)
         {
             RessourceId = ressourceid;
             RessourceModelString = ressourceModelString;
